+++ tests/test_integration/test_command.py
@@ -4,6 +4,7 @@ import subprocess
 from tests.lib.util import get_version_script, get_version_module, create_pyproject_toml, create_setup_py
 
 
+@pytest.mark.important
 def test_command_pyproject_toml(repo):
     create_pyproject_toml(repo)
 
@@ -11,6 +12,7 @@ def test_command_pyproject_toml(repo):
     assert get_version_script(repo) == "0.0.1"
 
 
+@pytest.mark.important
 def test_command_pyproject_toml_disabled(repo):
     create_pyproject_toml(repo, {"enabled": False})
 
@@ -21,6 +23,7 @@ def test_command_pyproject_toml_disabled(repo):
         get_version_script(repo)
 
 
+@pytest.mark.important
 def test_command_setup_py_fail(repo):
     create_setup_py(repo)